feat(unittests): add support for dynamic test harnesses
This patch adds support to the unittest framework to support
dynamic reconfiguration of test harnesses at run time, adding
flexibility to the tests.

It also replaces the build-time-linked test harnesses used
by the existing unit tests by dynamic ones and includes
some refactor on the slot buffer and the host test helpers to
ease the tests.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: Iabe9d0982d4fb6e78424fcd24a916324c7675697