refactor(xlat): Replace the xlat context macros with functions.
xlat_context module, in the xlat library, uses a number of macros
to allocate and initialize the translation tables as well as the
translation context structures.

Such macros make the xlat library harded to be unit-tested.

This patch removes these macros and repaces them with a set of
APIs that offer a similar functionality but can be used at build
time, making the unittests simpler.

The patch also adapts the rest of RMM to use such APIs.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I2fb486c0c5bc005446b09e3fef5e7de9bf0efda0