fix(unittest): remove dependency with enum buffer_slot in inst_helpers.h
inst_helpers.h has a dependency with enum buffer_slot as this is needed
by host_harness and the former includes the latter.

This patch breaks that dependency by removing a forward declaration
to `enum slot_buffer` inside host_harness.h, as C++ does not admit
forward declarations for enumerations.

Instead, it replaces the enumeration with an unsigned int, and replaces
the declaration of any functions in host_harness.h using the enumeration
by repacing such enumeration with an unsigned int.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: Ia26f771f5056047e9a61b2e02fb2294928c325d6