feat(xlat): make a distinction between transient and invalid descriptors
Currently, the xlat library treats transient and invalid descriptors
in the same way, printing the former as if they were invalid when
LOG_LEVEL is set to VERBOSE mode.

This patch introduces a new descriptor type, `TRANSIENT_DESC`, which,
whilst still an invalid descriptor, allows to differentiate it from
the rest of invalid descriptors so to print more accurate logs.

The patch uses one of the user available bits on the page/block
upper attributes area to mark a tte as transient, regardless of
whether the tte is valid or not.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I613117f76802de0d7c2c1c7546197fb4906f8ea2