fix(unittests): sysregs status might be inconsistent between runs
Current implementation of test_helpers_rmm_start() in the
test_helpers library performs an initialization of the RMM runtime
for the host build on the first call. Subsequent calls are just
ignored after the first initialization, so there is no need to
stop and deinitialize the runtime after every test to just
reinitialize everything again next time the runtime is needed.

This can cause a number of problems due to the sysregs left in
an undetermined status:

   - A test group which doesn't need to use the RMM runtime
     but is run after a test group which used such runtime
     might find some of the sysregs already initialized or in
     an inconsistent state, as the previous test group did not
     reset the "architectural" state of the host. This can cause
     inconsistencies on the tests.

   - If such test group then modifies or resets any sysreg used
     by the runtime, any test group run later which needs the
     runtime will fail, as any call to test_helpers_rmm_start()
     will be ignored and the sysregs will be left in an inconsisent
     state.

This patch adds support to store and restore a snapshot of the
RMM runtime for any test that needs to use it by invoking
test_helpers_rmm_start().

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I3efd4c881860cab6d9e02dc385c8bbcf868e7b2d