fix(unittests): sysregs status might be inconsistent between runs
Current implementation of test_helpers_rmm_start() in the
test_helpers library performs an initialization of the RMM runtime
for the host build on the first call. Subsequent calls are just
ignored after the first initialization.

This can cause a number of problems due to the sysregs left in
an undetermined status. For instance, a test using the RMM
runtime will leave the sysreg in an inconsistent state and cause
problems for the next test one. Hence it is essential for the
next test to reset to a previous known state as part of
initialization. Any test  using the RMM runtime will need to
restore the sysregs status and callbacks before running any new test.

This patch adds support to save the sysreg state after RMM
initialization in test_helpers_rmm_start() and restores this
state for any subsequent call to test_helpers_rmm_start().

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I3efd4c881860cab6d9e02dc385c8bbcf868e7b2d