refactor(plat/common): use arch specific shared buf va and pa
The fake_host arch needs to use the same address as the allocated
memory for access. The xlat_table address translation is not relevant
for fake_host. Hence this patch allows to abstract the va
when initializing the rmm-el3-ifc and the pa when creating the xlat tables
for shared buffer region. This also solves an issue wherein sometimes
the shared buf address collided with the dummy values used for mmap
regions on fake_host thus failing the initialization.

The patch also moves the dummy symbols for fake_host regions from
import_sym.h to plat_cmn_arch.h.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I70b9172573855e3fcf3508acabf3ee87ea19c57f