fix(lib/xlat): xlat_get_llt_from_va() returns the wrong base_va
xlat_get_llt_from_va() should populate xlat_llt_info struct with the
base VA mapped on the table being returned but instead, it populates
it with the context VA address.

This patch fixes that, as well as it renames some APIs and structures
in order to better reflect their purpose. In addition to that, it
also introduces some refactoring to improve performance related
to xlat_get_tte_ptr() API.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I1b16518d11b5427f3c5e02feb32e3ed9deb4e3e4