feat(rmm): check if map_addr is outside PAR
This patch makes map_unmap_ns() to check if map_addr
is outside PAR.
The handler used to rely on s2tte primitives to return
the error if the s2tte is not valid_ns.
This change aligns the eror codes of smc_rtt_map_unprotected()
and smc_rtt_unmap_unprotected() handlers and improves
readability of the source code.
This patch is based on
https://gerrit.oss.arm.com/c/fenimore/proto/rmm/+/239308

Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>
Change-Id: Ie60f9a579833e50ca5fe20be0d7251a54fc0f638