feat(rmm): rename and separate primitives
This patch describes the encodings of s2tte types, renames
"invalid_ns" to "unassigned_ns" and "valid_ns" to "assigned_ns"
to follow the naming pattern of other types of s2ttes.

The encoding of unassigned_ns (former invalid_ns) now differs from
the encoding of unassigned_empty so RMI_REALM_CREATE must do
the proper initialization of the starting level S2 TTs. So far the
corectness of the code was based on the assumption that the
encodings for unassigned hipas and empty ripas were zero.

This patch is based on:
https://gerrit.oss.arm.com/c/fenimore/proto/rmm/+/239509
https://gerrit.oss.arm.com/c/fenimore/proto/rmm/+/239309

Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>
Change-Id: Iea282baa4f367c88664406c86ad6b1569d4e89e6