feat(rmm): modify RMI/RSI commands logging
Number of input arguments and output values is
now encoded in 'rmi_type' of the handler.
Logging of RMI & RSI commands is modified to only
print supported input arguments.
For RMI commands output values are logged only
in case of RMI_SUCCESS return status.
Added logging of output values for RSI and PSCI
commands.
Minor changes to comply with MISRA-C and coding
style.

Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>
Change-Id: Ie53c6450495e49136f0a39752182363e7c93f09f