fix(fvp): Increase the RMM_MAX_GRANULES
Increase the size of RMM_MAX_GRANULES so that RMM can cater for
the case that the linux DT parsed by EL3 has not catered
for the 64 MB carveout. The carveout memory cannot be accessed
by Non Secure anyway, it is just that some build scripts
do not mark this carveout in DT and this then gets passed
down to RMM by EL3. There is no impact on RMM other than
slighly higher memory footprint. If NS access the memory, it
will fault.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I8148b4d48e40b17586d6d8f942328d4842bc28c6