feat(xlat): add support for FEAT_LPA2 to the stage 1 xlat library
The xlat library is able to work with and without FEAT_LPA2 supported
by the architecture.

This patch also hardcodes the shareability attribute for all
memory as ISH.

In order to maintain simplicity and make the review process
easier, this patch does not update the unittests nor any of the
host variants, hence they will not work on the current patch.

Unittests as well as the host variants will be fixed on a separate
patch.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I5501371c602f358a650b005eda2c2f9d076a4b33