feat(xlat): add support for FEAT_LPA2 to the stage 1 xlat library
The xlat library is able to work with and without FEAT_LPA2 supported
by the architecture.

This patch also hardcodes the shareability attribute for all
memory as ISH.

Whist this patch fixes the existing unittests for the xlat
library so they can run with the new changes, it does not introduce
any new testcases for FEAT_LPA2. These will be introduced in a
later patch.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I5501371c602f358a650b005eda2c2f9d076a4b33