fix(smccc): do not filter SMC calls unnecessarily
At the moment we have a big switch/case construct to handle and dispatch
SMCCC calls. Before we enter this switch/case, we do an explicit check
for accepted SMCCC numbers. This check is not only hard to read, it's
also tedious to maintain, since it needs to be in sync with the
switch cases.

Just let the compiler do the work, and issue the error message in the
default: case. This avoids accidental filtering, and also removes some
slightly dodgy code that upset code checkers.

Change-Id: Id25839cc73c2ff56d14226b9c262539f03205a11
Signed-off-by: Andre Przywara <andre.przywara@arm.com>