feat(s2tte): add support for FEAT_LPA2 to the s2tte library
For simplicity, when FEAT_LPA2 is supported by the architecture
for stage 2 translation, this patch enables it on the MMU
(e.g. by setting vtcr_el2.DS to '1'), regardless of whether LPA2
is enabled on a Realm by the Host or not.

During realm creation, RMM verifies the IPA size and start level
against LPA2 settings setup by the Host.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I1db0e7b9ce9ab5c67fd8a5027fbca27630ed85aa