docs: Update coding guidelines on use of 64 bit types
The RMM coding guidelines allow use of ``unsigned long``
, ``uint64_t`` and ``u_register_t`` for 64 bit types. Although
``u_register_t`` has its purposes, in the particular case of RMM
where the CPU is v9.2+ and is always 64-bit, it is a source of
non-uniformity in the codebase.

Also, update the guidelines for pointer type to give preference
to ``uintptr_t`` over ``void *``.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I377bf94b47f4c234e8777f6135e910bced115338