feat(lib/arch): add support for simd context
This change introduces SIMD context that holds the SIMD specific config
registers, vector registers and its status flags. This context can
represent the state of FPU/SVE registers of a Non-secure world or of a
Realm world.

The SIMD layer exposes APIs to runtime component of RMM to initialize,
save, restore, and switch SIMD context.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I0f78a3fdeeb13e0d05f5a51755962dfaa720ff68