fix(attest): point to correct repo for t_cose
The previous patch had updated t_cose to point to upstream
URL but since the SHA corresponds to fork which was used for
the development. This patch fixes the URL to point to fork
since it works better with the `repo` tool workflow used by
some projects.

The update to upstream URL will be done later when t_cose is
updated to an official release.

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: Id42f3d36dc9dc41dafbc5beb2dc01b6d0944cbce