feat(attest): adjust heap size depending on MAX_CPUS
The heap size needed for attestation depends on the number of possible
CPUs. Since the QEMU platform has a MAX_CPUS of 32 instead of the
default 16, 3 pages are not sufficient for the heap. Adjust the heap
size as a function of MAX_CPUS:

	  MAX_CPUS	| Heap size	| 4k Pages
	----------------+---------------+----------
	  16		| 10284		| 2.5
	  32		| 15744		| 3.8

Note that I used a simplified counter to measure the memory
requirements, instead of mbedtls_memory_buffer_alloc_status() which
seems unavailable to the RMM allocator.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Change-Id: I2d1e298c7732cc3ad28969dca41689afcf604680