feat(lib/arch): add SVE hint bit support
This change adds api simd_update_smc_sve_hint() to be used by SMC
handler like RMI or RSI to update the SVE hint passed in SMC function ID
to the SIMD context.

If the SVE hint is set for a SIMD context, then only Q0-Q31 registers
are saved and restored during the context switch.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I29adc61ac213a7cc0ca3567b1c80348ca752c3fa