feat(coverage): add a new rule to run coverage without dependencies
At the moment, 'run-coverage' rule on the RMM build system has a
dependency on 'run-unittests' rule, which first runs all the unittests
and then performs the coverage analysis.

This is useful to get a full picture of all the coverage done by the
whole unittests, but it can be cumbersome when developing a new set
of unittests as in this case we want to be able to run coverage
analysis ONLY to the tests under development in order to find out
how much has been covered so far.

This patch introduces a new rule, 'run-dev-coverage', which can be
used during unittest development to run coverage analysis without
running (nor building) anything before. The user will need to
build and run any test whose coverity wants to analyze manually
before invoking the 'run-dev-coverage' rule.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: Ibef3b4d265e13ea92e11bad9b05642703c855fea