fix(s2tt): RMM_DATA_CREATE_UNKNOWN sets RIPAS to RAM
According to the latest TF-RMM spec, RMM_DATA_CREATE_UNKNOWN does
not change the RIPAS of the s2tte where the data granule is mapped,
yet the current implementation sets it to RAM. This patch fixes
that by keeping the RIPAS state unaltered.

This patch also fixes a bug in smc_data_destroy in which the case
for HIPAS == ASSIGNED and RIPAS == DESTROYED was not taken into
account.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I98330dedca6951ddb29a19e8e38760c0b31f01b1