feat(rmm): suppress MISRA C:2012 Rule 10.4 violation
ULONG_MAX and ULLONG_MAX definitions in aarch64-none-elf\include\limits.h
produce MISRA C:2012 Rule 10.4 violations:
Essential type of the left hand operand "9223372036854775807UL" (signed)
is not the same as that of the right operand "2UL"(unsigned).
Essential type of the left hand operand "9223372036854775807ULL" (signed)
is not the same as that of the right operand "2ULL"(unsigned).

This patch replaces usage of ULONG_MAX and ULLONG_MAX
with ~(0UL) and ~(0ULL) respectively.

Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>
Change-Id: Ica9d641f862ca6f4d784b4e946b214a3f3d90b41