chore(s2tt): optimize s2tt component
This patch introduces a minor set of optimizations into the
s2tt and rtt components in order to improve performance.
Mainly, the optimizations consist on define some very simple
APIs as 'inline' to remove the calling/return overhead when
possible, as soon as other minor modifications.

This patch also fixes most of the MISRA warning already existing on the
modified files.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I8c96e76badeca560db316853ab1730df332a3474