fix(lib): clang-tidy NOLINT exclusions
The clang-tidy misc-unused-parameters check flags any function
parameters that are unused in the body of the function. However, this
check will give false positives if the parameter is used only in
assert() calls. This patch wraps the relevant function declarations
in NOLINT comments so that they are excluded from this check.

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: Ifb8288197a962fd2146d74d445141fe55ad61d8d