fix(fake_host): fix spinlock_acquire() implementation on the fake_host
Current implementation of the spinlock_acquire() harness on the
fake_host platform just sets up the spinlock to '1' and exits. This
means that if spinlock_acquire() is called whilst 'acquired', the call
will no make any effect and the code will continue to execute.

In order for the execution to be consistent with the acquire semantics,
this patch introduces an assert that will interrupt the execution flow
if spinlock_acquire() is invoked over an already acquired spinlock.
For the host_test variant, the unittest will be able to catch the
assertion and react according to the desired test configuration.
Conversely, the host_build variant will abort execution if an already
locked spinlock tries to get locked again. Note that this should
never happen on the latter variant as the fake host is always emulated
in a single CPU on the host, and therefore race conditions should
never happen.

This patch also fixes a couple of bugs on the granule unittests
discovered after the reimplementation of the spinlock_acquire()
harness.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: If784edcc725ebbc7b4e0bbe2fbdb87acff47af4a