feat(tools/cbmc): add script to check CBMC results
This commit adds a python script that compares the CBMC results to the
baseline summaries that are committed in the source tree. If there is a
significant difference, the script returns an error.

A call to the script is added to the CheckCBMC.cmake file. If the script
returns error, the script output is printed on the console, however the
build is not interrupted with an error.

Change-Id: I50e7b9a570c1d30e03def630c6541d63536bd1c1
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>