fix(tools): add warnings count for commit message check
Currently the CheckPatch.cmake script outputs the total number of errors
and warnings from the commit message checker. However, the commit
message checker script does not actually count the number of warnings,
so the warning count is always zero.

This patch updates checkcommitmsg.py and CheckPatch.cmake such that the
number of warnings is correctly counted and printed.

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: Ie6855192ac01e81eb43ed4ea7d224cf0583a678b