test(lib/arch): demonstrate usage of callbacks for SIMD unit tests
This patch demonstrates how the callback framework for SIMD unit tests
can be used. A callback is installed to count the number of times the
fpu_save_registers() function is called. This count is then checked in a
unit test to verify the control flow of the simd_context_switch()
function.

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: I57626cdbdd36222fea223adf2b2470b078ce6c11