refactor(s2tt): pass the realm_s2_context to all the s2tt APIs
This patch adds the realm_s2_context structure as parameter on every
single API on the s2tt library, regardless of whether it is needed or
not. This allows the library to get information on the realm for which
the call is intended to be used so that the information can be used for
further features to be added to the library.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I467a42ccd43809a5cddf82c13cf7e0911b969da8