feat(plat/arm): Introduce the common arm platform
This patch introduced the common arm platform which can be used
by any compatible hardware. This allows reuse of code across multiple
platforms and possibly a single binary which can run on these platforms.
This depends on the MAX_CPU and RMM_MAX_GRANULES and RMM_MAX_SIZE can
encompass the superset of the values.

Signed-off-by: Harry Moulton <harry.moulton@arm.com>
Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I69ceea101595f0e8cac81bd9e4d8754d2b1edbc3