feat(unittests): add unittests for the s2tt library
This patch adds unittests to cover a subset of the APIs of the
s2tt library. Unittests for the rest of the APIs will be provided
on a later patch.

This patch also fixes some minor bugs on the s2tt library detected
whilst developing the unittests. The bugs were mostly related to
incomplete checks.

The patch also removes some dead code detected on the s2tt library
while running the tests. The dead code is mostly redundant asserts.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I9052b5a19d1f1a842c4ebd244c74b24ae1f7fa25