feat(lib/queue): Add simple circular buffer
Add a simple circular buffer implementation that can have a power of 2
capacity. The intent for this simple implementation is to be used for
cases where queue space is fixed at compile time. The implementation
itself is not multi-thread safe and the expectation is that a locking
mechanism outside of the library is used before making any function call
in this library. Therefore, it is not meant for extreme performance
sensitive paths.

Change-Id: Ic2186dacea2cf2455fa635c9071f5f333494bb95
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>