feat(shrinkwrap): add shrinkwrap configurations for RMM
This patch adds a shrinkwrap configuration to build and run RMM TFTF
tests using the current local (or, alternatively, a different one)
RMM repository. The configuration included in this patch can be
found at:

	- <RMM_ROOT>/shrinkwrap/configs/rmm-tftf.yaml

In addition to the above conriguration, this patch includes the
following overlays, that can be used in any needed combination:

	- <RMM_ROOT>/shrinkwrap/common/layers/tftf.yaml
	- <RMM_ROOT>/shrinkwrap/common/layers/rmm.yaml
	- <RMM_ROOT>/shrinkwrap/common/layers/rmm-debug.yaml
	- <RMM_ROOT>/shrinkwrap/common/layers/model-enable-cache.yaml
	- <RMM_ROOT>/shrinkwrap/common/layers/model-enable-lpa2.yaml
	- <RMM_ROOT>/shrinkwrap/common/layers/model-wait-debugger.yaml

Of all the above overlays, the first two are already used by the
configuration, whereas the rest of overlays can be used as needed
to test any particular functionality.

This patch also fixes a minor bug on the documentation in which a
license table on the README was not properly rendered.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I8ccd243d8580ad6ddb4b146674f55c8d6b5b1df9