feat(planes): define per-plane stage 2 translation contexts
This patch replaces the current s2tt_context structure on the
realm descriptor with an array of structures, one per plane. It
also adds some APIs to access the different contexts.

The patch only adds support for the primary plane and the
feature registers are left untouched, so there is no effective
change in functionality.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I9c8d5318a68c26f35dcf986918335609b284fabe