fix(runtime/rmi): make INJECT_SEA mutually exclusive to EMUL_MMIO
The invocation of mmio_emulation() and sea_inj() functions need to be
mutually exclusive during schedule of a REC. Prior to the fix, both the
cases are allowed to be satisfied at the same time which is incorrect.

This patch does the fix such that if INJECT_SEA is set, then it skips
actions for EMUL_MMIO flag. This is  because the conditions for the
INJECT_SEA  flag having an effect (Data Abort at Unprotected IPA) are a
superset of those for RMI_EMULATED_MMIO (Data Abort at Unprotected IPA
and access was an emulatable read)

Signed-off-by: Soby Mathew <soby.mathew@arm.com>
Change-Id: I427d1d8bbbe8ffbde84bfc8f05b528fc4d473edf