refactor(lib): restore COMPILER_ASSERT for CBMC
COMPILER_ASSERT is a useful tool for CBMC to detect issues that can
lead to assert failures. A previous commit completely turned off
COMPILER_ASSERT because some alignment and structure size requirements
are not fulfilled by the CBMC configuration.

This commit introduces new macros for asserting on structure field
offset and structure size, and implements them as always passing for
CBMC. However other COMPILER_ASSERTs are turned on for CBMC.

Change-Id: I20cd80bd5ca86ee92407c92e26ac2c99698a3e05
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>