feat(lib/slot_buf): add slot for REC
Add a new slot buffer for a REC granule to be mapped. This is used to
map an additional REC, that may not be the currently running REC, to be
able to write into it, when HES is used for attestation. Also add slots
for the Aux granules which host the crypto context, that may need to be
accessed when HES is used for attestation.

Change-Id: Ib9b36f46967f738c7a2ad5ee7553701af1e99279
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>