feat(runtime/core): check HES queues
When the RMM uses the HES for attestation, check on every entry and
exit to and from the RMM, the request queues and response buffer. If the
response buffer has a response, write the response to the appropriate
REC. If there are new requests pending, send it to the HES for signing.
The design always attempts to push requests to EL3 unless it returns an
error indicating there is no more queue space. This design allows EL3 to
have a single one request one response interface, that suits simple EL3
and HES designs with mailbox, and also allows for more complex EL3
designs where requests and responses to HES can be queued in EL3 as well
for the HES to process them in a platform specific way.
The appropriate hes queue functions are called after handling an NS SMC,
on every entry into realms and in the attestation loop when no
interrupts are pending for realms.

Change-Id: Ia6b6941d49e36d07733eea9303d5634c2dde0093
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>