feat(planes): handle S2 access permissions in s2tt.c primitives
This patch adds support for AP in RTTEs with RIPAS = RAM and
assumes that other RTTEs (RIPAS == EMPTY/DESTROYED) do not
maintain AP value.

In addition, this patch moves the RIPAS field to avoid clashes
with the AP field and expands its size from 2 to 3 bits to
support future expansion.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I34c39b6238f8de155b8e92bc8d386865c524e053