feat(planes): handle S2 access permissions in s2tt.c primitives
This patch stores Access Permissions in al S2 RTTES and
preserves the Access Permission values on all APIs except
for rmi_rtt_destroy() which sets the APs to the default value.

In addition, this patch moves the RIPAS field to avoid clashes
with the AP field and expands its size from 2 to 3 bits to
support future expansion.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I34c39b6238f8de155b8e92bc8d386865c524e053