feat(lib/attestation): add HES context
Add context for enabling attestation signing to use the HES. In this
patch, the crypto context that is forwarded to the t_cose library is
replaced by the context as defined in t_cose HES adapter. This patch
also adds an init function in the attestation library to be called by
the rumtime to initialize the context. Replace initialization of
attestation state using these functions.

Change-Id: I9ffe989c75112e002d567dd8c5bf3f8814f10d5c
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>