feat(runtime/core): add support for EL3 attestation queue
This patch adds code to enable the use of a EL3 queue to queue
attestation requests. When the RMM uses the t_cose adapter to sign
attestation requests using EL3, the added code initializes call backs to
take the CBOR encoded requests and hashes to be signed and enqueues it
in EL3.
The code added also provides functions that can be called to pull
responses from EL3.

Change-Id: Icbc1139f9c5fee5e1a229a964fdda291a332a8b8
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>