fix(runtime/rmi): fix deadlock
When aux granules are being locked during REC create, if there is an
error on one of the granules, we free them. However, in the free
function, the lock is acquired again before releasing the lock. This
patch fixes the issue by first unlocking the granules, before calling
the function to free them.

Change-Id: Id0b4764e1ae134d44caa383b924251deb1d2f44a
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>