fix(lib/s2tt): fix checks on s2tte_get_ripas()
Current implementation of s2tte_get_ripas() does not accept a valid
descriptor as it asserts on the bit fields wrongly.

This patch removes the incorrect the assert() and comment for valid
descriptor case. In case the descriptor is valid, the RIPAS is RAM.

Note that this function is not being invoked for a valid descriptor
currently in RMM and hence this assert does not fire during run.

Closes #21.

Signed-off-by: Javier Almansa Sobrino <javier.almansasobrino@arm.com>
Change-Id: I7b28c7eb63b1303aaa491983457771be957405f6