fix(lib/xlat): remove broadcast invalidates
The RMM uses high high va ranges to map and unmap buffers during its
operation. These ranges are per core private mappings. This patch
optimizes the unmapping sequence to enforce ordering and TLB
invalidation only on the local core and avoid the broadcast invalidate
which can be expensive on large core systems.
This patch also removes the incorrectly used COMPILER_BARRIER
during the call to unmapping APIs, and also fixes the translation
table write to use store release, to ensure memory ops before the
write to the TTE are observed before the write to the TTE.

Change-Id: I2b39b39ade9a0bbe59b2ff7bd98a3664c63c3623
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>