refactor(runtime/core): optimize REC map unmap
This patch optimizes the function that pulls responses from the HES
queue to skip lock, map and unmap of the rec granule corresponding
to the response, if it is the same as the currently running rec on
the PE.
This patch also removes the versioning of the structures for the
request and responses to the HES since the EL3<->RMM interface is
already versioned and accounts for it. This patch also removes the
memset to zero on the shared buf to avoid unnecessary writes.

Change-Id: Ie68001112f7e5f99b379b1fc1b55adb02558675a
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>