refactor(lib/attestation): change return values cca token
Function attest_cca_token_create() in the Attestation component
did not follow the same behavior for the return value as the other
ones. This patch modifies the return values for the function so it
now returns an enum attest_token_err_t like the other ones in the
attestation_token module. As the function previously returned the
length of the CCA top-level as the return value, this value has
been converted to an output argument.

New error values were added to account for the behavior of this
attest_cca_token_create(). Now it returns ATTEST_TOKEN_ERR_SUCCESS
in case of success or the proper error value otherwise.

Change-Id: I9979d96c5d2863490819aec47a6eef1c4e9a45af
Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>