fix(runtime): clear the attestation data after token creation
The patch f362213ee1 moved attestation state management inside the
component. But the buffer used by the caller for attest data was not
reset on successful token creation which allowed another
ATTEST_TOKEN_CONTINUE call to go through. This patch does the needed reset of the buffer and solves the issue.

Change-Id: If24994b29d2288fcdd85755cc53a23fcf7fa3882
Signed-off-by: Soby Mathew <soby.mathew@arm.com>