feat(lib/rmm_el3_ifc): add RMM_EL3_FEATURES call
Add support for RMM_EL3_FEATURES call. The call is used to
discover features and calls supported by EL3, for example,
ability to use EL3 for attestation signing and the corresponding
EL3 interfaces. The features call itself requires the RMM-EL3
interface to be at version 0.3 to be called. EL3 implements a
feature register with a fixed definition, to be able to discover
supported features.

Change-Id: I0ac007bf8224818884d07de97009248b7a62781a
Signed-off-by: Raghu Krishnamurthy <raghupathyk@nvidia.com>