Handle realm metadata and sealing
This commit adds handling of realm metadata and sealing key derivation.

There is a dedicated RMI for provisioning the metadata to the realm,
and verification of its signature and validity.

The realm metadata mechanism is optional. You can ran realms without
privided metadata, although during the derivation process, one
won't be able to derive stable keys immune to realm updates.

Signed-off-by: Piotr Sawicki <p.sawicki2@samsung.com>